backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[google-genai]>=1.6.0",
    "crewai-tools>=1.6.0",
    "fastapi>=0.120.4",
    "googlenews>=1.6.15",
    "langchain-openai>=1.1.0",
    "loguru>=0.7.3",
    "praw>=7.8.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "uvicorn>=0.38.0",
    "apscheduler>=3.11.1",
]
